Move Temporal polyfill management to temporal.js
